# perf(encoder) add small encoder optimizations