@brettlangdon 2021-06-30
Closing this as it was just exploration.